refactor: Remove Relay Classic support
