+++ src/QuerySubscription.js
