+++ src/Resolver.js
