--- src/classic/index.js
@@ -1,2 +1,0 @@
-export RelayRouteRenderer from './RelayRouteRenderer';
-export Resolver from './Resolver';
