--- src/modern/index.js
@@ -1,1 +1,0 @@
-export Resolver from './Resolver';
