--- test/classic/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    ["env", {
-      "loose": true
-    }],
-    "stage-1",
-    "react"
-  ],
-  "plugins": [
-    ["relay", { "schema": "test/fixtures/schema.graphql" }]
-  ]
-}
