+++ test/helpers.js
@@ -1,9 +1,9 @@
 import { createFetch } from 'relay-local-schema';
 import { Environment, Network, RecordSource, Store } from 'relay-runtime';
 
-import { Resolver } from '../../src';
+import { Resolver } from '../src';
 
-import schema from '../fixtures/schema';
+import schema from './fixtures/schema';
 
 export function createFakeFetch() {
   return createFetch({ schema });
