+++ test/render-warnings.test.js
@@ -7,7 +7,7 @@ import ReactDOMServer from 'react-dom/server';
 import { graphql } from 'react-relay';
 import warning from 'warning';
 
-import { Resolver } from '../../src';
+import { Resolver } from '../src';
 
 import { createEnvironment } from './helpers';
